+++ packages/eas-cli/src/update/configure.ts
@@ -9,13 +9,14 @@ import { ExpoGraphqlClient } from '../commandUtils/context/contextUtils/createGr
 import { AppPlatform } from '../graphql/generated';
 import Log, { learnMore } from '../log';
 import { RequestedPlatform, appPlatformDisplayNames } from '../platform';
-import { createOrModifyExpoConfigAsync } from '../project/expoConfig';
+import { createOrModifyExpoConfigAsync, isUsingStaticExpoConfig } from '../project/expoConfig';
 import {
   installExpoUpdatesAsync,
   isExpoUpdatesInstalledAsDevDependency,
   isExpoUpdatesInstalledOrAvailable,
 } from '../project/projectUtils';
 import { resolveWorkflowPerPlatformAsync } from '../project/workflow';
+import { confirmAsync } from '../prompts';
 import { syncUpdatesConfigurationAsync as syncAndroidUpdatesConfigurationAsync } from './android/UpdatesModule';
 import { syncUpdatesConfigurationAsync as syncIosUpdatesConfigurationAsync } from './ios/UpdatesModule';
 
@@ -324,7 +325,8 @@ export async function ensureEASUpdateIsConfiguredInEasJsonAsync(projectDir: stri
 /**
  * Make sure EAS Update is fully configured in the current project.
  * This goes over a checklist and performs the following checks or changes:
- *   - Enure the `expo-updates` package is currently installed.
+ *   - Ensure `updates.useClassicUpdates` (SDK 49) is not set in the app config
+ *   - Ensure the `expo-updates` package is currently installed.
  *   - Ensure `app.json` is configured for EAS Updates
  *     - Sets `runtimeVersion` if not set
  *     - Sets `updates.url` if not set
@@ -333,7 +335,7 @@ export async function ensureEASUpdateIsConfiguredInEasJsonAsync(projectDir: stri
 export async function ensureEASUpdateIsConfiguredAsync(
   graphqlClient: ExpoGraphqlClient,
   {
-    exp: expWithoutUpdates,
+    exp: expMaybeWithoutUpdates,
     projectId,
     projectDir,
     platform,
@@ -344,9 +346,17 @@ export async function ensureEASUpdateIsConfiguredAsync(
     platform: RequestedPlatform | null;
   }
 ): Promise<void> {
+  // EAS Update and SDK 49's "useClassicUpdates" option are mutually exclusive
+  if (expMaybeWithoutUpdates.updates?.useClassicUpdates) {
+    expMaybeWithoutUpdates = await ensureUseClassicUpdatesIsRemovedAsync({
+      exp: expMaybeWithoutUpdates,
+      projectDir,
+    });
+  }
+
   const hasExpoUpdates = isExpoUpdatesInstalledOrAvailable(
     projectDir,
-    expWithoutUpdates.sdkVersion
+    expMaybeWithoutUpdates.sdkVersion
   );
   const hasExpoUpdatesInDevDependencies = isExpoUpdatesInstalledAsDevDependency(projectDir);
   if (!hasExpoUpdates && !hasExpoUpdatesInDevDependencies) {
@@ -367,7 +377,7 @@ export async function ensureEASUpdateIsConfiguredAsync(
   const workflows = await resolveWorkflowPerPlatformAsync(projectDir);
   const { projectChanged, exp: expWithUpdates } =
     await ensureEASUpdatesIsConfiguredInExpoConfigAsync({
-      exp: expWithoutUpdates,
+      exp: expMaybeWithoutUpdates,
       projectDir,
       projectId,
       platform,
@@ -391,4 +401,50 @@ export async function ensureEASUpdateIsConfiguredAsync(
     );
     Log.newLine();
   }
+}
+
+export async function ensureUseClassicUpdatesIsRemovedAsync({
+  exp: expMaybeWithoutUpdates,
+  projectDir,
+}: {
+  exp: ExpoConfig;
+  projectDir: string;
+}): Promise<ExpoConfig> {
+  if (!isUsingStaticExpoConfig(projectDir)) {
+    throw new Error(
+      `Your app config sets "updates.useClassicUpdates" but EAS Update does not support classic updates. Remove "useClassicUpdates" from your app config and run this command again.`
+    );
+  }
+
+  const shouldEditConfig = await confirmAsync({
+    message: `Your app config sets "updates.useClassicUpdates" but EAS Update does not support classic updates. Remove "updates.useClassicUpdates" from your app config?`,
+  });
+  if (!shouldEditConfig) {
+    throw new Error(
+      `Manually remove "updates.useClassicUpdates" from your app config and run this command again.`
+    );
+  }
+
+  const editedExpoConfig = mergeExpoConfig(expMaybeWithoutUpdates, {
+    updates: { useClassicUpdates: undefined },
+  }) as ExpoConfig;
+  const result = await createOrModifyExpoConfigAsync(projectDir, editedExpoConfig);
+
+  switch (result.type) {
+    case 'success':
+      Log.withTick(`Removed "updates.useClassicUpdates"`);
+      expMaybeWithoutUpdates = editedExpoConfig;
+      break;
+
+    case 'warn':
+    case 'fail':
+      throw new Error(result.message);
+
+    default:
+      throw new Error(
+        `Unexpected result type "${result.type}" received when modifying the project config.`
+      );
+  }
+
+  return editedExpoConfig;
 }